2023/march/yolo-v1/requirements.txt
absl-py==1.4.0
albumentations==1.3.0
black @ file:///opt/conda/conda-bld/black_1660237809219/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click @ file:///tmp/build/80754af9/click_1646056706450/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
docopt==0.6.2
fonttools==4.25.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
grpcio==1.53.0
idna==3.4
imageio==2.27.0
imgaug==0.4.0
joblib==1.2.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lazy_loader==0.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib @ file:///croot/matplotlib-suite_1678723872360/work
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
networkx==3.0
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.5.3
pathspec @ file:///croot/pathspec_1674681560568/work
Pillow==9.4.0
pipreqs==0.4.11
platformdirs @ file:///opt/conda/conda-bld/platformdirs_1662711380096/work
ply==3.11
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1671697431263/work
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.3.21
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarg==0.1.9
